Started DownloadServiceImpl refactor
