+++ ultrasonic/src/main/java/org/moire/ultrasonic/provider/UltraSonicAppWidgetProvider.java
@@ -19,6 +19,7 @@ import org.moire.ultrasonic.activity.MainActivity;
 import org.moire.ultrasonic.domain.MusicDirectory;
 import org.moire.ultrasonic.service.DownloadService;
 import org.moire.ultrasonic.service.DownloadServiceImpl;
+import org.moire.ultrasonic.service.MediaPlayerService;
 import org.moire.ultrasonic.util.FileUtil;
 
 public class UltraSonicAppWidgetProvider extends AppWidgetProvider
@@ -69,11 +70,11 @@ public class UltraSonicAppWidgetProvider extends AppWidgetProvider
 	/**
 	 * Handle a change notification coming over from {@link DownloadService}
 	 */
-	public void notifyChange(Context context, DownloadService service, boolean playing, boolean setAlbum)
+	public void notifyChange(Context context, MusicDirectory.Entry currentSong, boolean playing, boolean setAlbum)
 	{
 		if (hasInstances(context))
 		{
-			performUpdate(context, service, null, playing, setAlbum);
+			performUpdate(context, currentSong, null, playing, setAlbum);
 		}
 	}
 
@@ -96,15 +97,14 @@ public class UltraSonicAppWidgetProvider extends AppWidgetProvider
 	/**
 	 * Update all active widget instances by pushing changes
 	 */
-	private void performUpdate(Context context, DownloadService service, int[] appWidgetIds, boolean playing, boolean setAlbum)
+	private void performUpdate(Context context, MusicDirectory.Entry currentSong, int[] appWidgetIds, boolean playing, boolean setAlbum)
 	{
 		final Resources res = context.getResources();
 		final RemoteViews views = new RemoteViews(context.getPackageName(), this.layoutId);
 
-		MusicDirectory.Entry currentPlaying = service.getCurrentPlaying() == null ? null : service.getCurrentPlaying().getSong();
-		String title = currentPlaying == null ? null : currentPlaying.getTitle();
-		String artist = currentPlaying == null ? null : currentPlaying.getArtist();
-		String album = currentPlaying == null ? null : currentPlaying.getAlbum();
+		String title = currentSong == null ? null : currentSong.getTitle();
+		String artist = currentSong == null ? null : currentSong.getArtist();
+		String album = currentSong == null ? null : currentSong.getAlbum();
 		CharSequence errorState = null;
 
 		// Show error message?
@@ -117,7 +117,7 @@ public class UltraSonicAppWidgetProvider extends AppWidgetProvider
 		{
 			errorState = res.getText(R.string.widget_sdcard_missing);
 		}
-		else if (currentPlaying == null)
+		else if (currentSong == null)
 		{
 			errorState = res.getText(R.string.widget_initial_text);
 		}
@@ -157,7 +157,7 @@ public class UltraSonicAppWidgetProvider extends AppWidgetProvider
 		// Set the cover art
 		try
 		{
-			Bitmap bitmap = currentPlaying == null ? null : FileUtil.getAlbumArtBitmap(context, currentPlaying, 240, true);
+			Bitmap bitmap = currentSong == null ? null : FileUtil.getAlbumArtBitmap(context, currentSong, 240, true);
 
 			if (bitmap == null)
 			{
@@ -176,7 +176,7 @@ public class UltraSonicAppWidgetProvider extends AppWidgetProvider
 		}
 
 		// Link actions buttons to intents
-		linkButtons(context, views, currentPlaying != null);
+		linkButtons(context, views, currentSong != null);
 
 		pushUpdate(context, appWidgetIds, views);
 	}
@@ -200,19 +200,19 @@ public class UltraSonicAppWidgetProvider extends AppWidgetProvider
 
 		// Emulate media button clicks.
 		intent = new Intent("1");
-		intent.setComponent(new ComponentName(context, DownloadServiceImpl.class));
+		intent.setComponent(new ComponentName(context, MediaPlayerService.class));
 		intent.putExtra(Intent.EXTRA_KEY_EVENT, new KeyEvent(KeyEvent.ACTION_DOWN, KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE));
 		pendingIntent = PendingIntent.getService(context, 0, intent, 0);
 		views.setOnClickPendingIntent(R.id.control_play, pendingIntent);
 
 		intent = new Intent("2");  // Use a unique action name to ensure a different PendingIntent to be created.
-		intent.setComponent(new ComponentName(context, DownloadServiceImpl.class));
+		intent.setComponent(new ComponentName(context, MediaPlayerService.class));
 		intent.putExtra(Intent.EXTRA_KEY_EVENT, new KeyEvent(KeyEvent.ACTION_DOWN, KeyEvent.KEYCODE_MEDIA_NEXT));
 		pendingIntent = PendingIntent.getService(context, 0, intent, 0);
 		views.setOnClickPendingIntent(R.id.control_next, pendingIntent);
 
 		intent = new Intent("3");  // Use a unique action name to ensure a different PendingIntent to be created.
-		intent.setComponent(new ComponentName(context, DownloadServiceImpl.class));
+		intent.setComponent(new ComponentName(context, MediaPlayerService.class));
 		intent.putExtra(Intent.EXTRA_KEY_EVENT, new KeyEvent(KeyEvent.ACTION_DOWN, KeyEvent.KEYCODE_MEDIA_PREVIOUS));
 		pendingIntent = PendingIntent.getService(context, 0, intent, 0);
 		views.setOnClickPendingIntent(R.id.control_previous, pendingIntent);
