+++ ultrasonic/src/main/java/org/moire/ultrasonic/receiver/MediaButtonIntentReceiver.java
@@ -25,9 +25,9 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Parcelable;
 import android.util.Log;
-import android.view.KeyEvent;
 
 import org.moire.ultrasonic.service.DownloadServiceImpl;
+import org.moire.ultrasonic.service.MediaPlayerService;
 import org.moire.ultrasonic.util.Util;
 
 /**
@@ -57,41 +57,22 @@ public class MediaButtonIntentReceiver extends BroadcastReceiver
 			Parcelable event = (Parcelable) extras.get(Intent.EXTRA_KEY_EVENT);
 			Log.i(TAG, "Got MEDIA_BUTTON key event: " + event);
 
-			Intent serviceIntent = new Intent(context, DownloadServiceImpl.class);
+			Intent serviceIntent = new Intent(context, MediaPlayerService.class);
 			serviceIntent.putExtra(Intent.EXTRA_KEY_EVENT, event);
 
 			try
 			{
-				context.startService(serviceIntent);
-			}
-			catch (IllegalStateException exception)
-			{
-				Log.i(TAG, "MediaButtonIntentReceiver couldn't start DownloadServiceImpl because the application was in the background.");
-				if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
-				{
-					KeyEvent keyEvent = (KeyEvent) event;
-					if (keyEvent.getAction() == KeyEvent.ACTION_DOWN && keyEvent.getRepeatCount() == 0)
-					{
-						int keyCode = keyEvent.getKeyCode();
-						if (keyCode == KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE ||
-								keyCode == KeyEvent.KEYCODE_HEADSETHOOK ||
-								keyCode == KeyEvent.KEYCODE_MEDIA_PLAY)
-						{
-							// TODO: The only time it is OK to start DownloadServiceImpl as a foreground service is when we now it will display its notification.
-							// When DownloadServiceImpl is refactored to a proper foreground service, this can be removed.
-							context.startForegroundService(serviceIntent);
-							Log.i(TAG, "MediaButtonIntentReceiver started DownloadServiceImpl as foreground service");
-						}
-					}
-				}
-			}
+				if (DownloadServiceImpl.getInstance() == null) new DownloadServiceImpl(context);
+				DownloadServiceImpl.getInstance().onCommand(serviceIntent);
 
-			try
-			{
 				if (isOrderedBroadcast())
 				{
 					abortBroadcast();
 				}
+			}
+			catch (IllegalStateException exception)
+			{
+				Log.w(TAG, "MediaButtonIntentReceiver couldn't start DownloadServiceImpl because the application was in the background.");
 			}
 			catch (Exception x)
 			{
