+++ ultrasonic/src/main/java/org/moire/ultrasonic/service/DownloadFile.java
@@ -441,10 +441,7 @@ public class DownloadFile
 
 				new CacheCleaner(context, DownloadServiceImpl.getInstance()).cleanSpace();
 
-				if (DownloadServiceImpl.getInstance() != null)
-				{
-					((DownloadServiceImpl) DownloadServiceImpl.getInstance()).checkDownloads();
-				}
+				MediaPlayerService.checkDownloads(context);
 			}
 		}
