+++ ultrasonic/src/main/java/org/moire/ultrasonic/service/Supplier.java
@@ -1,0 +1,6 @@
+package org.moire.ultrasonic.service;
+
+public abstract class Supplier<T>
+{
+    public abstract T get();
+}
